add kategori permission

+++ app/Http/Controllers/Dashboard/Admin/CategoryController.php
@@ -9,26 +9,31 @@ class CategoryController extends Controller
 {
     public function index()
     {
+        $this->authorize('lihat kategori', \App\Models\Category::class);
+
         return view('pages.dashboard.admin.categories');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('tambah kategori', \App\Models\Category::class);
+
         $request->validate([
             'name' => 'required|string',
-            'slug' => 'required|string|unique:categories,slug',
-            'image' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-            'link' => 'nullable|string',
-            'type' => 'required|in:internal,external',
+            //'image' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            //'link' => 'nullable|string',
+            //'type' => 'required|in:internal,external',
         ]);
 
-        $data = $request->all();
+       // if ($request->hasFile('image')) {
+           // $data['image'] = $request->file('image')->store('categories', 'public');
+       // }
 
-        if ($request->hasFile('image')) {
-            $data['image'] = $request->file('image')->store('categories', 'public');
-        }
-
-        $category = \App\Models\Category::create($data);
+        $category = \App\Models\Category::create([
+            'name' => $request->name,
+            'slug' => \Illuminate\Support\Str::slug($request->name),
+            'link' => "#",
+            ]);
 
         return redirect()->route('dashboard.admin.categories.index')->with('success', 'Category created successfully');
     }
@@ -59,6 +64,8 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('hapus kategori', \App\Models\Category::class);
+
         $category = \App\Models\Category::findOrFail($id);
         $category->delete();
